requirements.txt
py-cord==2.6.1
colorama~=0.4.6
aiohttp~=3.10.11
bs4~=0.0.2
art~=6.2
requests~=2.32.0

keyboard~=0.13.5
Flask~=3.0.2
pillow~=10.3.0
numpy~=1.26.4
matplotlib~=3.8.4
beautifulsoup4~=4.12.3
pypresence~=4.3.0
scipy~=1.14.1
psutil~=6.1.1